requirements.txt
Flask-JWT-Extended
Flask-Smorest
passlib
marshmallow
python-dotenv
gunicorn
mongoengine